requirements.txt
gitpython
clinner